slightly faster rand(big(a):big(b))
For generating `x` in a `0:m` BigInt range, instead of uniformly
randomizing all the limbs of `x` and rejecting when out of range,
we generate the highest limb `hx` of `x` in `0:hm` where `hm`
is the highest limb of `m`. Before the introduction of
`SamplerRangeNDL`, this would have changed nothing, as the predecessor
`SamplerRangeFast` was itself using rejection sampling by generating
first a given number of random bits. But with NDL, this can speeds-up
`BigInt` generation by very roughly 10%, as in the general case
almost no "rejection" will happen.