sometimes faster `rand(::Xoshiro, ::UnitRange{Int128})`
For Int128/UInt128 ranges, the default sampler provided by `Random`
is `SamplerRangeFast`; till now, this was optimized for a version
of MersenneTwister which produced natively 52 bits of entropy,
so if the range length was between 2^52 and 2^104, two 52-bits
words were generated (and used with rejection sampling).
But when the RNG can produce a 64-bits word in a fast way,
only one such word needs to be generated for a length
between 2^52 and 2^64: in this range, the new `rand` method
for `SamplerRangeFast` is about 35% faster.
Also, `MersenneTwister` can now generate `UInt64` fast, but the two
methods are equally fast, so it was not changed for that RNG.